Search plant name against GBIF, KNMS and POWO names backbone

Perform a fuzzy search against three names list to get best match and POWO taxonomic status

name_search(name)

Arguments

name
A scientific plant species name. Better results can be obtained when author is included e.g. Poa annua L.

Value

Returns a data frame with ...

Examples

# single name search
name_search("Poa annua L.")

#> Warning: 'format.knms_match' is deprecated.
#> Use 'tidy.knms_match' instead.
#> See help("Deprecated")

#> Warning: 'format.powo_taxon' is deprecated.
#> Use 'tidy.powo_taxon' instead.
#> See help("Deprecated")

#> # A tibble: 1 x 9
#>   taxonomicStatus submitted  matched ipni_id matched_record  searchName usageKey
#>   <chr>           <chr>      <lgl>   <chr>   <chr>           <chr>         <int>
#> 1 Accepted        Poa annua~ TRUE    320035~ Poa annua Sp. ~ Poa annua~  2704179
#> # ... with 2 more variables: scientificName <chr>, confidence <int>

# Or, search multiple names using purrr::map_dfr
names <- c("Poa annua L.", "Welwitschia mirabilis Hook.f.")
names_out <- purrr::map_dfr(names, name_search)

#> Warning: 'format.knms_match' is deprecated.
#> Use 'tidy.knms_match' instead.
#> See help("Deprecated")

#> Warning: 'format.powo_taxon' is deprecated.
#> Use 'tidy.powo_taxon' instead.
#> See help("Deprecated")

#> Warning: 'format.knms_match' is deprecated.
#> Use 'tidy.knms_match' instead.
#> See help("Deprecated")

#> Warning: 'format.powo_taxon' is deprecated.
#> Use 'tidy.powo_taxon' instead.
#> See help("Deprecated")