Search plant name against GBIF names backbone

This function uses the rgbif package https://docs.ropensci.org/rgbif/index.html to query a scientific plant name against the GBIF names backbone - see https://www.gbif.org/. It restricts the result to species rank constrained to kingdom = 'Plantae'. This function uses the rgbif::name_backbone_verbose function.

name_search_gbif(name)

Arguments

name
A scientific plant species name. Better results can be obtained when author is included e.g. Poa annua L.

Value

Returns a data frame with initial search term searchName, GBIF taxon key usageKey, GBIF scientific name scientificName, and a measure of confidence in the match. When there is no match it returns a value of "no_match" under the confidence field.

Examples

# single name search
name_search_gbif("Poa annua L.")

#> # A tibble: 4 x 4
#>   searchName   usageKey scientificName             confidence
#>   <chr>           <int> <chr>                           <int>
#> 1 Poa annua L.  2704179 Poa annua L.                      100
#> 2 Poa annua L.  8422205 Poa annua Cham. & Schltdl.         83
#> 3 Poa annua L. 11019629 Poa annua L.Sp.Pl                  78
#> 4 Poa annua L.  7730008 Poa annua Steud., 1854             78

# Or, search multiple names using purrr::map_dfr
names <- c("Poa annua L.", "Welwitschia mirabilis Hook.f.")

#names_out <- purrr::map_dfr(names, name_search_gbif)